+++ lib/simple_form/form_builder.rb
@@ -17,19 +17,20 @@ module SimpleForm
     extend MapType
     include SimpleForm::Inputs
 
-    map_type :text,                                       to: SimpleForm::Inputs::TextInput
-    map_type :file,                                       to: SimpleForm::Inputs::FileInput
-    map_type :string, :email, :search, :tel, :url, :uuid, to: SimpleForm::Inputs::StringInput
-    map_type :password,                                   to: SimpleForm::Inputs::PasswordInput
-    map_type :integer, :decimal, :float,                  to: SimpleForm::Inputs::NumericInput
-    map_type :range,                                      to: SimpleForm::Inputs::RangeInput
-    map_type :check_boxes,                                to: SimpleForm::Inputs::CollectionCheckBoxesInput
-    map_type :radio_buttons,                              to: SimpleForm::Inputs::CollectionRadioButtonsInput
-    map_type :select,                                     to: SimpleForm::Inputs::CollectionSelectInput
-    map_type :grouped_select,                             to: SimpleForm::Inputs::GroupedCollectionSelectInput
-    map_type :date, :time, :datetime,                     to: SimpleForm::Inputs::DateTimeInput
-    map_type :country, :time_zone,                        to: SimpleForm::Inputs::PriorityInput
-    map_type :boolean,                                    to: SimpleForm::Inputs::BooleanInput
+    map_type :text,                                to: SimpleForm::Inputs::TextInput
+    map_type :file,                                to: SimpleForm::Inputs::FileInput
+    map_type :string, :email, :search, :tel, :url, to: SimpleForm::Inputs::StringInput
+    map_type :citext, :uuid,                       to: SimpleForm::Inputs::StringInput
+    map_type :password,                            to: SimpleForm::Inputs::PasswordInput
+    map_type :integer, :decimal, :float,           to: SimpleForm::Inputs::NumericInput
+    map_type :range,                               to: SimpleForm::Inputs::RangeInput
+    map_type :check_boxes,                         to: SimpleForm::Inputs::CollectionCheckBoxesInput
+    map_type :radio_buttons,                       to: SimpleForm::Inputs::CollectionRadioButtonsInput
+    map_type :select,                              to: SimpleForm::Inputs::CollectionSelectInput
+    map_type :grouped_select,                      to: SimpleForm::Inputs::GroupedCollectionSelectInput
+    map_type :date, :time, :datetime,              to: SimpleForm::Inputs::DateTimeInput
+    map_type :country, :time_zone,                 to: SimpleForm::Inputs::PriorityInput
+    map_type :boolean,                             to: SimpleForm::Inputs::BooleanInput
 
     def self.discovery_cache
       @discovery_cache ||= {}