+++ lib/simple_form/form_builder.rb
@@ -136,11 +136,13 @@ module SimpleForm
     #        name="user[name]" type="text" value="Carlos" />
     #
     def input_field(attribute_name, options={})
+      components = (wrapper.components & ATTRIBUTE_COMPONENTS)
+
       options = options.dup
-      options[:input_html] = options.except(:as, :collection, :label_method, :value_method, *ATTRIBUTE_COMPONENTS)
+      options[:input_html] = options.except(:as, :collection, :label_method, :value_method, *components)
       options = @defaults.deep_dup.deep_merge(options) if @defaults
 
-      SimpleForm::Wrappers::Root.new(ATTRIBUTE_COMPONENTS + [:input], wrapper: false).render find_input(attribute_name, options)
+      SimpleForm::Wrappers::Root.new(components + [:input], wrapper: false).render find_input(attribute_name, options)
     end
 
     # Helper for dealing with association selects/radios, generating the