+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.0.3
+
+### bug fix
+  * Fix XSS vulnerability on error components.
+
 ## 3.0.2
 
 ### enhancements